kernel : allow compiling with -O2

+++ kernel/arch/x86_64/cpu/gdt.c
@@ -42,7 +42,7 @@ void init_gdt(void){
 
 	//tss take two entries
 	kernel->arch.gdt[5] = create_gdt_segement((uint64_t)&kernel->arch.tss & 0xFFFFFFFF,sizeof(TSS) - 1,0x89,0);
-	uint32_t tss_addressH = ((uint64_t)&kernel->arch.tss >> 32) & 0xFFFFFFFF;
+	uint64_t tss_addressH = ((uint64_t)&kernel->arch.tss >> 32) & 0xFFFFFFFF;
 	kernel->arch.gdt[6] = *(gdt_segment *)&tss_addressH;
 
 	//create the GDTR and load it so the GDT is actually used
